+++ classes/Integration/Form/ContactForm7.php
@@ -87,11 +87,15 @@ class PUM_Integration_Form_ContactForm7 extends PUM_Abstract_Integration_Form {
 	 * @param WPCF7_ContactForm $cfdata
 	 */
 	public function on_success( $cfdata ) {
-		/**
-		 * @see pum_integrated_form_submission
-		 */
+
+		if ( ! self::should_process_submission() ) {
+			return;
+		}
+		$popup_id = self::get_popup_id();
+		self::increase_conversion( $popup_id );
+
 		pum_integrated_form_submission( [
-			'popup_id'      => isset( $_REQUEST['pum_form_popup_id'] ) && absint( $_REQUEST['pum_form_popup_id'] ) > 0 ? absint( $_REQUEST['pum_form_popup_id'] ) : false,
+			'popup_id'      => $popup_id,
 			'form_provider' => $this->key,
 			'form_id'       => $cfdata->id(),
 		] );
@@ -105,11 +109,6 @@ class PUM_Integration_Form_ContactForm7 extends PUM_Abstract_Integration_Form {
 	 * @return array
 	 */
 	public function custom_scripts( $js = [] ) {
-		$js[ $this->key ] = [
-			'content'  => file_get_contents( Popup_Maker::$DIR . 'assets/js/pum-integration-' . $this->key . PUM_Site_Assets::$suffix . '.js' ),
-			'priority' => 8,
-		];
-
 		return $js;
 	}