+++ classes/Upsell.php
@@ -1,8 +1,13 @@
 <?php
 
+/**
+ * Handles displaying promotional text throughout plugin UI
+ */
 class PUM_Upsell {
 
-
+	/**
+	 * Hooks any needed methods
+	 */
 	public static function init() {
 		add_filter( 'views_edit-popup', array( __CLASS__, 'addon_tabs' ), 10, 1 );
 		add_filter( 'views_edit-popup_theme', array( __CLASS__, 'addon_tabs' ), 10, 1 );
@@ -11,8 +16,9 @@ class PUM_Upsell {
 	}
 
 	/**
-	 * @param array $tabs
+	 * Adds messages throughout Popup Settings UI
 	 *
+	 * @param array $tabs The tabs/fields for popup settings.
 	 * @return array
 	 */
 	public static function popup_promotional_fields( $tabs = array() ) {
@@ -44,8 +50,9 @@ class PUM_Upsell {
 	}
 
 	/**
-	 * @param array $tabs
+	 * Adds messages throughout Popup Theme UI
 	 *
+	 * @param array $tabs The tabs/fields for popup theme.
 	 * @return array
 	 */
 	public static function theme_promotional_fields( $tabs = array() ) {
@@ -71,9 +78,7 @@ class PUM_Upsell {
 	 * When the Popup or Popup Theme list table loads, call the function to view our tabs.
 	 *
 	 * @since 1.8.0
-	 *
-	 * @param $views
-	 *
+	 * @param array $views An array of available list table views.
 	 * @return mixed
 	 */
 	public static function addon_tabs( $views ) {
@@ -93,39 +98,35 @@ class PUM_Upsell {
 		$theme_labels = PUM_Types::post_type_labels( __( 'Popup Theme', 'popup-maker' ), __( 'Popup Themes', 'popup-maker' ) );
 
 		?>
+		<style>
+			.wrap h1.wp-heading-inline + a.page-title-action {
+				display: none;
+			}
 
+			.edit-php.post-type-popup .wrap .nav-tab-wrapper .page-title-action, .edit-php.post-type-popup_theme .wrap .nav-tab-wrapper .page-title-action, .popup_page_pum-extensions .wrap .nav-tab-wrapper .page-title-action {
+				top: 7px;
+				margin-left: 5px
+			}
 
-        <style>
-            .wrap h1.wp-heading-inline + a.page-title-action {
-                display: none;
-            }
-
-            .edit-php.post-type-popup .wrap .nav-tab-wrapper .page-title-action, .edit-php.post-type-popup_theme .wrap .nav-tab-wrapper .page-title-action, .popup_page_pum-extensions .wrap .nav-tab-wrapper .page-title-action {
-                top: 7px;
-                margin-left: 5px
-            }
-
-            @media only screen and (min-width: 0px) and (max-width: 783px) {
-                .edit-php.post-type-popup .wrap .nav-tab-wrapper .page-title-action, .edit-php.post-type-popup_theme .wrap .nav-tab-wrapper .page-title-action, .popup_page_pum-extensions .wrap .nav-tab-wrapper .page-title-action {
-                    display: none !important
-                }
-            }
-        </style>
-
-
-        <h2 class="nav-tab-wrapper">
+			@media only screen and (min-width: 0px) and (max-width: 783px) {
+				.edit-php.post-type-popup .wrap .nav-tab-wrapper .page-title-action, .edit-php.post-type-popup_theme .wrap .nav-tab-wrapper .page-title-action, .popup_page_pum-extensions .wrap .nav-tab-wrapper .page-title-action {
+					display: none !important
+				}
+			}
+		</style>
+		<nav class="nav-tab-wrapper">
 			<?php
 			$tabs = array(
 				'popups'       => array(
-					'name' => $popup_labels['name'],
+					'name' => esc_html( $popup_labels['name'] ),
 					'url'  => admin_url( 'edit.php?post_type=popup' ),
 				),
 				'themes'       => array(
-					'name' => $theme_labels['name'],
+					'name' => esc_html( $theme_labels['name'] ),
 					'url'  => admin_url( 'edit.php?post_type=popup_theme' ),
 				),
 				'integrations' => array(
-					'name' => __( 'Extensions and Integrations', 'popup-maker' ) . PUM_Admin_Extend::append_unseen_count(),
+					'name' => esc_html__( 'Extensions and Integrations', 'popup-maker' ) . PUM_Admin_Extend::append_unseen_count(),
 					'url'  => admin_url( 'edit.php?post_type=popup&page=pum-extensions&view=integrations' ),
 				),
 			);
@@ -134,9 +135,10 @@ class PUM_Upsell {
 
 			$active_tab = false;
 
+			// Calculate which tab is currently active.
 			if ( isset( $_GET['page'] ) && $_GET['page'] === 'pum-extensions' ) {
 				$active_tab = 'integrations';
-			} else if ( ! isset( $_GET['page'] ) && isset( $_GET['post_type'] ) ) {
+			} elseif ( ! isset( $_GET['page'] ) && isset( $_GET['post_type'] ) ) {
 				switch ( $_GET['post_type'] ) {
 					case 'popup':
 						$active_tab = 'popups';
@@ -147,24 +149,25 @@ class PUM_Upsell {
 				}
 			}
 
+			// Add each tab, marking the current one as active.
 			foreach ( $tabs as $tab_id => $tab ) {
-
-				$active = $active_tab == $tab_id ? ' nav-tab-active' : '';
-
-				echo '<a href="' . esc_url( $tab['url'] ) . '" class="nav-tab' . $active . '">';
-				echo $tab['name'];
-				echo '</a>';
+				$active = $active_tab === $tab_id ? ' nav-tab-active' : '';
+				?>
+				<a href="<?php echo esc_url( $tab['url'] ); ?>" class="nav-tab<?php echo esc_attr( $active ); ?>">
+					<?php echo $tab['name']; ?>
+				</a>
+				<?php
 			}
 			?>
 
-            <a href="<?php echo admin_url( 'post-new.php?post_type=popup' ); ?>" class="page-title-action">
-				<?php echo $popup_labels['add_new_item']; ?>
-            </a>
+			<a href="<?php echo esc_url( admin_url( 'post-new.php?post_type=popup' ) ); ?>" class="page-title-action">
+				<?php echo esc_html( $popup_labels['add_new_item'] ); ?>
+			</a>
 
-            <a href="<?php echo admin_url( 'post-new.php?post_type=popup_theme' ); ?>" class="page-title-action">
-				<?php echo $theme_labels['add_new_item']; ?>
-            </a>
-        </h2>
+			<a href="<?php echo esc_url( admin_url( 'post-new.php?post_type=popup_theme' ) ); ?>" class="page-title-action">
+				<?php echo esc_html( $theme_labels['add_new_item'] ); ?>
+			</a>
+		</nav>
 		<?php
 	}