+++ popup-maker.php
@@ -3,7 +3,7 @@
  * Plugin Name:  Popup Maker
  * Plugin URI:   https://wppopupmaker.com/?utm_campaign=PluginInfo&utm_source=plugin-header&utm_medium=plugin-uri
  * Description:  Easily create & style popups with any content. Theme editor to quickly style your popups. Add forms, social media boxes, videos & more.
- * Version:      1.12.0
+ * Version:      1.13.0
  * Author:       Popup Maker
  * Author URI:   https://wppopupmaker.com/?utm_campaign=PluginInfo&utm_source=plugin-header&utm_medium=author-uri
  * License:      GPL2 or later
@@ -93,7 +93,7 @@ class Popup_Maker {
 	/**
 	 * @var string Plugin Version
 	 */
-	public static $VER = '1.12.0';
+	public static $VER = '1.13.0';
 
 	/**
 	 * @var int DB Version
@@ -113,7 +113,7 @@ class Popup_Maker {
 	/**
 	 * @var string
 	 */
-	public static $MIN_WP_VER = '3.6';
+	public static $MIN_WP_VER = '4.9';
 
 	/**
 	 * @var string Plugin URL